Fix checkSignThenDecrypt_RSA1()'s payload stream consumer: Detect EOS also if net-file-size has been reached and signal EOS.
We don't want to rely on passed 'is_final', since its crucial to call `aead->finish(data);`
on the last chunk to get rid of potantial padding bytes.

Also we like to signal to the 'feeder loop' to end transmission if satisfied and to close stream.
This relaxes requirements to the 'feeder loop' and renders the implementation more robust.